Move FSE-specific functions outside of main plugin file
